+++ .gitignore
@@ -3,3 +3,6 @@ node_modules/
 
 # archives
 package-lock.json
+
+# environnement
+.env
