+++ package.json
@@ -14,15 +14,16 @@
     "nodemon": "^2.0.22"
   },
   "dependencies": {
+    "bcrypt": "^5.1.0",
     "bootstrap": "^5.2.1",
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "jquery": "^3.6.1",
-    "popper.js": "^1.16.1",
-    "socket.io": "^4.5.2",
-    "socket.io-client": "^4.5.2",
-    "bcrypt": "^5.1.0",
     "jsonwebtoken": "^9.0.0",
     "mongoose": "^7.0.4",
-    "mongoose-unique-validator": "^4.0.0"
+    "mongoose-unique-validator": "^4.0.0",
+    "popper.js": "^1.16.1",
+    "socket.io": "^4.5.2",
+    "socket.io-client": "^4.5.2"
   }
 }
