+++ server.js
@@ -47,8 +47,8 @@ server.on('listening', () => {
 const io = require('socket.io')(server, {
   cors: {origin : '*'}
 });
-
-app.set('io', io);
+// initialize my socketio module and pass it the io instance
+require('./socket')(io);
 
 server.listen(port);
